ai-agent/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Langchain
langchain>=0.1.0,<0.2.0
langchain-openai>=0.0.2,<0.1.0
langchain-community>=0.0.10,<0.1.0

# Tavily Search
tavily-python>=0.3.0

# Database
mysql-connector-python==8.2.0
pymysql==1.1.0
sqlalchemy==2.0.23

# Environment and Configuration
python-dotenv==1.0.0

# HTTP and API
httpx==0.25.2
requests==2.31.0

# Date and Time
python-dateutil==2.8.2

# CORS
fastapi-cors==0.0.6

# OpenAI (for Langchain)
openai>=1.6.1,<2.0.0

